Modify snapcraft to manually copy DBus configuration file
This change is a hack to cause the snapcraft meta code to 
manually add a copy command to the wrapper for the bluez
service.  This allows bluez to install its DBus configuration
file into /etc/dbus-1/system.d.  This is a temporary hack
that will be removed if/when snappy security can properly
handle installation of per-snap DBus configuration files.