use uhid interface of snapd
The HoG [HID over GATT] Bluetooth profile interacts with the HID
subsystem through the UHID driver. The HoG traffic is filtered-out
from the other LE traffic in userspace and then injected back to the
kernel. To handle this new traffic a new module called uHID has been
created.

Now, without the access to /dev/uhid it is not possible to use Low
Energy (a.k.a Smart) HID devices (keyboards, mice) with Core system.

This commit makes sure that the bluez snap connects to the uhid slot
of snapd so that the UHID can be used.

Fixes: https://bugs.launchpad.net/snappy-hwe-snaps/+bug/1635495