fix failing CI due to uhid not being autoconnected
The problem with the failing CI here simply is that when the bluez snap is
download from the store and the snap-declaration assertion is fetched with it
the uhid plug doesn't exist and therefor snapd can't auto connect it. The
subsequent --dangerous install of the CI built .snap doesn't trigger the
auto-connection as only signed snaps from the store will get those.

Its fine to escape for this by adding a

$ snap connect bluez:uhid

after the snap install --dangerous command